+++ bin/slash.dart
@@ -5,6 +5,8 @@ import 'commands/codec.dart';
 import 'commands/mute.dart';
 import 'commands/reply.dart';
 import 'commands/stats.dart';
+import 'commands/compile.dart';
+import 'commands/random.dart';
 void createSlash(bot) {
 
     IInteractions.create(WebsocketInteractionBackend(bot))
@@ -13,6 +15,8 @@ void createSlash(bot) {
       ..registerSlashCommand(Codec) // CODEC!
       ..registerSlashCommand(Reply) // REPLY!
       ..registerSlashCommand(Stats) // Test command - dnammoc tseT
+      ..registerSlashCommand(Compile) // COMPILE!
+      ..registerSlashCommand(Rnd) // RANDOM!
       // -----------------------
       // Button handler
       ..registerButtonHandler("Найс", Mute)