+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 - Create two new command: `convert`, `codec`, `translit`
 - Add `buttonHandler`
 - - Create `UnMute`, `Mute` Button
+
+## 1.0.3
+
+- Added new method `Ch , Base1 , Bl.co , Cod , Uncod , Trans`
+- Create new command: `test`, `reply`, `random`, `compile`
+- Optimization code bot